package/lapack: introduce BR2_PACKAGE_LAPACK_ARCH_SUPPORTS
This makes it easier for packages that depend on lapack to get
their dependencies correct.

The !uClibc dependency is also not sufficient: indeed, musl too does not
provide _fpu_control; only glibc does. This is the same situation as for
clapack. Add a comment about this, to mirror clapack.

Since the !glibc dependency only exists for PowerPC, treat it as
an architecture dependency.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
[yann.morin.1998@free.fr:
  - fpu_control depends on glibc, not on !uclibc
  - add or update comments accordingly
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>