package/kmsxx: remove erroneous use of static-libc option
Currently, for a static-only build, we pass -Dstatic-libc=true. The
reason for that was not recorded when the package was converted to
meson.

The -Dstatic-libc=true option, despite its name, is not about linking
statically against libc, but against libgcc and libstdc++.

In Buildroot, we forcibly pass -static when calling the compiler and
linker, so everything is already linked statically.

For a shared build, -Dstatic-libc=false has no effect at all.

Drop this option altogether, as it is not needed, and is confusing.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
[yann.morin.1998@free.fr: expand commit log]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>