configs/acmesystems_acqua_a5_*: bump at91bootstrap3 version
AT91Bootstrap 3.10.2 has a bug in its Makefile:[1] it gives the
-nostartfiles flag to the ld command, although this is a gcc flag, not
an ld flag. The bug happens to be harmless when using binutils 2.35.
However, the ld shipped with binutils 2.36 is more strict in its option
parsing, and fails with the following error:

    .../output/host/bin/arm-buildroot-linux-uclibcgnueabihf-ld:
    Error: unable to disambiguate: -nostartfiles
    (did you mean --nostartfiles ?)

The bug was fixed upstream in commit 3207586a.[2] However, since there
has been no tagged 3.x release since then, we just upgrade to the
current head of the branch at91bootstrap-3.x.

Fixes:

  https://gitlab.com/buildroot.org/buildroot/-/jobs/1478738145
  https://gitlab.com/buildroot.org/buildroot/-/jobs/1478738152

[1] https://github.com/linux4sam/at91bootstrap/issues/127
[2] https://github.com/linux4sam/at91bootstrap/commit/3207586a79aba527bbf20560930089b800f191ac

Signed-off-by: Edgar Bonet <bonet@grenoble.cnrs.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>