od: Use a unique S03mountdata.sh init.d script
Instead of using one script per OpenDingux flavour, use one script that
will handle all flavours.

This has the benefit that a rs90 build, for instance, can be booted
from a SD card, and the data partition of the SD card will be used
properly.

It also supports using filesystems that aren't built-in but exported as
external modules. This allows the kernel to provide different
filesystems that can be used for the data partition (e.g. EXT4 or F2FS)
without clobbering the kernel with drivers that may not be used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>