package/nginx: override endianness
Though several cross-compilation patches exist in buildroot's nginx
package dir they do not seem to address endianness.

The test program generated by the configure script compiles but fails
to run (as it is built for another architecture) but the script does
not distinguish between the failure to run the program and an
indication of certain endianness.  As such the fallback of big-endian
is used.  This setting then causes http2 headers (anything not in the
static dictionary) to come out as undecipherable trash on 64bit
targets (see ngx_http_v2_huff_encode_buf()).

This commit includes a patch to the configure script to allow a
`--force-endianness=big|little` flag as well as setting that flag in
buildroot's package makefile.

Signed-off-by: Nevo Hed <nhed+buildroot@starry.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit e205b5ec1819d44f95c1886c5004be0eeabf56b3)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>