board/raspberrypicm4io: enable RTC and USB overlays
The CM4 IO Board is equipped with an on-board RTC connected over I2C,
and has USB 2.0 host ports.

This commit loads the appropriate overlays to enable both. The USB 2.0
interface is by default disabled on CM4 to reduce the power
consumption so it needs to be explicitly enabled on the CM4 IO Board
to use the USB ports.

Signed-off-by: Peter Seiderer <ps.report@gmx.net>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>