fs/iso9660: fix non-reproducible builds
In commit 179ae068eb70 (fs/iso9660: add support to Grub EFI bootloader
in the image), we did a last-minute change when applying the patch, on
the flawed assumption that the commands were run in a standard Makefile
rule.

However, for filesystems, most commands are run in a script (so they run
under fakeroot). As such, we can't silence the commands with the usual
Makefile '@' construct.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>