CI: Rework CI script
Build the installer as part of the 'rs90' job. This allows to reduce the
build time by about 20 minutes, since the 'installer' build only needs
the rs90 toolchain to build.

It also simplifies the script a lot, since there is no need to download
and install the rs90 toolchain since we just built it.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>