od: Add support for joysticks translator in SDL2
The Joystick Translator is a feature developed for SDL2.
See: https://github.com/libsdl-org/SDL/pull/4448

This commit allows applications using SDL2 to use the Gamepad API
without having to do anything. The capabilities of the gamepad exposed
to the application depends on the hardware the application is running
on.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>