package/python-urwid: depend on python3
Although the package claims to support both python2 and python3,
there are some places which contain python3-only code (mainly
asyncio-related).
We can try to tweak some of the files in order to make it work with
python2, but it doesn't worth the effort since python2 isn't maintained
for a long time.

Update the reverse recursive dependencies accordingly.

Fixes:
 - http://autobuild.buildroot.net/results/356c6b09566c6020dc320eb89da791581571cb7f/

Signed-off-by: Asaf Kahlon <asafka7@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>