package/luajit: back to mainline and bump to 05f1984e
Because upstream was inactive, we switched to the moonjit fork in commit
2ca0accc. Now, the moonjit fork is abandonned and Mike Pall resumed
their activity on LuaJIT. Therefore, we can move back to mainline.

There is no new upstream release yet, so take a commit from git. The
official site https://luajit.org/git/luajit.git doesn't offer tarball
downloads, so use the github mirror instead.

The moonjit part is no longer in the COPYRIGHT file (obviously). Also,
compared to the 2.0.5 release, the URLs in the file have been changes to
https. Therefore, the hash is updated.

Also, switch the hash file to two spaces.

Signed-off-by: Francois Perrad <francois.perrad@gadz.org>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>