package/dtbocfg: depends on kernel
Fix the following build failure raised since the addition of the package
in commit 9fc7b49ab193da55b0a72d27b5e0ea33245b4df0:

Makefile:587: *** linux is in the dependency chain of dtbocfg that has added it to its _DEPENDENCIES variable without selecting it or depending on it from Config.in.  Stop.

Fixes:
 - http://autobuild.buildroot.org/results/449a731c0a1394c28f519e341ea145cb42af012d

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Acked-by: Herve Codina <herve.codina@bootlin.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>