package/nftables: disable debug build
nftbles' --with-debug option enables both debugging symbols (with -g)
and debugging macros (with -DDEBUG).

The former are globally driven by the BR2_ENABLE_DEBUG and BR2_DEBUG_n
set of options, and enforced in our toolchain wrapper; the latter should
be driven by the BR2_ENABLE_RUNTIME_DEBUG, but nftables only uses it to
activate an assert-like macro in its json output generator, so we can
well live without that.

So we just unconditionally disable debug in nftables.

Signed-off-by: Peter Seiderer <ps.report@gmx.net>
[yann.morin.1998@free.fr: write a commit log]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>