package/openjdk{-bin}: bump version to 17.0.1+12
OpenJDK 17 is a new LTS release, which leaves the Buildroot Config option of
"LTS" and "LATEST" as a misnomer because both 11 and 17 are LTS releases.

There are two options in this case:

1) Remove "LATEST" and update OpenJDK 11 to 17, and only support 17.
2) Change "LTS" to "11" and "LATEST" to "17" and only support the latest 2 LTS
OpenJDK releases.

After some discussion with Thomas Petazzoni and Peter Korsgaard, and testing,
option 2 is the best course of action for a few reasons:

  - OpenJDK 11 and 17 have very long support cycles:
  - OpenJDK 11 has two years of Active and five years of security support left.
  - OpenJDK 17 has five years of Active and ten years of security support left.
  - Both OpenJDK versions build with the same parameters.
  - The maintenance cost of both versions is meager.
  - Both versions pass tests.package.test_openjdk without issue.

Changes:
  - Change BR2_OPENJDK_VERSION_LATEST -> BR2_OPENJDK_VERSION_17
  - Change BR2_OPENJDK_VERSION_LTS -> BR2_OPENJDK_VERSION_11
Signed-off-by: Adam Duskett <aduskett@gmail.com>
[Peter: add Config.in.legacy, use BR2_PACKAGE_OPENJDK_ prefix]
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
(cherry picked from commit a610bf996798949f047c1f84f5617be26c3682c6)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>