package/pipewire-media-session: new package
The pipewire-media-session has been moved into a separate package
which pipewire attempts to download via meson wraps if
session-managers are enabled, since we don't support meson wraps
we need to disable session-managers in the pipewire package and
create a new pipewire-media-session package.

We also need to add a patch removing an invalid session-managers
option check from pipewire-media-session.

There is an alsa with-module-sets option in pipewire-media-session,
however at the moment alsa is an unconditional dependency so
we need to always enable it and select alsa-lib.

Note: it was previously an option of the pipewire package, named
media-session, so the symbol was BR2_PACKAGE_PIPEWIRE_MEDIA_SESSION.
Now it is a package name pipewire-media-session, so the symbol is still
BR2_PACKAGE_PIPEWIRE_MEDIA_SESSION. So there is no legacy entry needed,
and users previously using that option will now get the package. Sneaky.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
[yann.morin.1998@free.fr:
  - extend commit log with legacy handling
  - first unconditional CONF_OPTS uses =, not +=; fits on one line
  - slight reorder in variables assignments
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>