package/mpd: update to version 0.23.y
This version adds support for openmpt, pipewire and snapcast, which are
disabled for now.

It also adds support for a choice between pupnp and npupnp for the UPnP
database. Therefore, the UPNP option is converted into three-value
choice in Config.in. Legacy handling of the BR2_PACKAGE_MPD_UPNP makes
sure that upnp remains selected.

Signed-off-by: Andreas Ziegler <br015@umbiko.net>
[Arnout: fix legacy handling; fix snapcast option]
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>