package/linux-firmware: add option for Cypress FullMAC
Since 2016, Cypress is in charge of supporting old Broadcom wireless
chips and producing new ones. Newer firmwares for now come with
a different license and have anoter location in a package.

Now, most of Broadcom FullMac firmwares in brcm/ directory should now
be just symlinks to the corresponding files in cypress/ directory.

Respectively, it also makes sence to autoselect corresponding CYPRESS
options if user selectes BRCM wireless stuff.

This patch was created as a result of investigation the issue with
broken wireless support for such popular boards as Raspberry Pi Zero W
and BananaPi M2 Zero.

Signed-off-by: Uladzimir Bely <wiselord1983@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>