package/python-numpy: simplify PYTHON_NUMPY_FIXUP_NPY_PKG_CONFIG_FILES
The PYTHON_NUMPY_FIXUP_NPY_PKG_CONFIG_FILES macro was introduced in
commit
2426002d6fc81d0b023e3f3b3ee4e94ab29438cb ("package/python-numpy: fixup
npymath.ini for numpy extensions").

However, it turns out that by using $(PYTHON3_PATH), we can simplify
the intermediate variables, to the point where they are no longer
needed. Not having the intermediate variables makes the logic a bit
easier to understand.

Also, the logic is changed to use a single sed expression, which we
pass to $(SED).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>