Config.in.legacy: remove BR2_PACKAGE_SYSTEMD_TMPFILES
When the BR2_PACKAGE_SYSTEMD_TMPFILES was removed, it was added to
Config.in. The idea was to warn users who use the symbol in their
br2-external.

However, most users don't have a br2-external that uses this symbol.
Since the symbol defaults to y when systemd is enabled, they'll get this
legacy warning - which is completely meaningless for them since nothing
changes.

The annoyance of the many is worse than the possible unexpected change
for the few. And anyone actually relying on BR2_PACKAGE_SYSTEMD_TMPFILES
will probably easily discover the issue in testing (e.g. because the
tmpfiles conf is missing in the target).

Therefore, remove BR2_PACKAGE_SYSTEMD_TMPFILES from Config.in.legacy.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>