board/spike/riscv64/start.sh: add buildroot host dir in PATH
commit 0d0f84d2004b04bc93d5ceb58c86640f190e5d78
added the missing host-dtc dependency to riscv-isa-sim.

The spike simulator calls the dtc binary at its startup. The host dtc
command needs to be in the PATH at that time.

This commit add the buildroot host binary directory into the PATH of the
start.sh helper script. It make sure spike will use the buildroot dtc
version. This commit fixes this start.sh script on host not providing
the dtc command. Since the "spike" binary is now in the PATH, the
binary name is also simplified.

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>