package/polkit: bump to fix CVE-2021-4034
We are currently using the fork by Adam, with support for duktape as a
JS engine. But upstream has finally merged that just a day ago. Woohoo!

Between the fork we were using and upstream, there are very little
changes, mostly centered around:
  - translations
  - buildsystem
  - duktape
  - CVE-2021-4034

So, we just switch to using the HEAD of the repo: it has not much more
than the two important changes: duktape and the CVE fix.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Marek Belisko <marek.belisko@open-nandra.com>
Cc: Adam Duskett <aduskett@gmail.com>
Cc: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
(cherry picked from commit 4aecb964aff92e5cc956b2d3b700ac7841041463)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>