package/systemd: always delete hwdb
Up to now, the source files of the hwdb was only cleaned if systemd-hwdb
was used. However, those files should only be used by the hwdb anyway --
otherwise we shouldn't clean them even when the hwdb is enabled.
Therefore, remove them even if hwdb is disabled.

hwdb files can be put in two places: /usr/lib/udev, or /etc/udev. The
former is where systemd-compliant packages would install them, but the
latter is where they were originally placed. In systemd-land, /etc
should only contain local information, so ideally it should be empty in
the rootfs.

The reality, however, is that many packages still install files in /etc.
Thus, there are probably also packages that install hwdb files in
/etc/udev.

Therefore, clean the hwdb files both in /etc/udev and in /usr/lib/udev.

Probably, we should do this cleaning even when systemd is not selected,
since non-systemd packages may install files there which are not
relevant without systemd-hwdb (or eudev). However, that's a bigger
change which needs a separate patch (and a bit more consideration).

While we're at it, get rid of the HOST_EUDEV_SYSCONFDIR variable. This
is a leftover from when udev was still extracted from the systemd
package (i.e. before eudev became a separate package).

Signed-off-by: Norbert Lange <nolange79@gmail.com>
[Arnout: split in two separate commits]
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>