package/ghostscript: disable tesseract without threads
ghostscript embeds a local copy of tesseract-ocr which is enabled by
default since version 9.53.0 and
https://git.ghostscript.com/?p=ghostpdl.git;a=commit;h=7322ef87546c55b29056c1d96d8bdbb7f3ba53ef

Since version 9.55.0 and
https://git.ghostscript.com/?p=ghostpdl.git;a=commit;h=5459f7aca1ded3b7cf96c339a2419f418aa406c7,
ghostscript expects that tesseract is disabled if threads are not
available resulting in the following build failure since commit
7a4944569ce83b0b7b0cabd6e8aa184b8cb72a22:

configure: error: Threading disabled or not available. Tesseract OCR relies on threading. Rerun configure with "--without-tesseract" to exclude OCR from the build

So disable tesseract if threads are not available. It should be noted
that there is no way to tell ghostscript to use a system provided
tesseract-ocr package.

Fixes:
 - http://autobuild.buildroot.org/results/fd35a47b2a77cf57c243ec1c88802b53ddca8f52

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>