DEVELOPERS: drop libqrtr-glib from Yegor Yefremov
Commit 73ce754531b5e45bd42b56c58718fbacc754406f in master added  the
libqrtr-glib package to Yegor Yefremov.

This commit was backported as fad154253deb5b5399ea52a19ed6608dcd00ee4d
in the LTS 2021.02.x branch, but this branch does not have the
libqrtr-glib package, as it was added after 2021.02.

Due to this, the DEVELOPERS file is bogus in 2021.02.x, which causes
the following warning when running pkg-stats:

WARNING: 'package/libqrtr-glib/' doesn't match any file

This commit fixes this by dropping the bogus line from the DEVELOPERS
file.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>