package/wpe/wpebackend-rdk: move feature selection to kconfig
Using the kconfig fragment to select which backend features to choose
for a given configuration is more flexible (and easier to understand)
than writing a bunch of conditionals in the Make fragment. This also
fixes issues related to dependencies not added correctly, e.g. before
picking Raspberry Pi support would result in libinput support also being
chosen, but the "libinput" package was not added to _DEPENDENCIES.