falcosecurity-libs: add new package
This is a dependency of newer sysdig. It contains the driver, and also a
few userspace components. The latter however are not meant to be
installed in the sysroot; instead, the whole thing is meant to be
included directly in the build of the project using it. Changing things
so it does work in the normal way of installing to the sysroot turns out
to be pretty complicated.

Basically, falcosecurity-libs is just a component of sysdig. It's
defined as a separate package only because that's an easier way to
download it than defining extra download and extract commands in sysdig
itself. For this reason, it's defined as a blind option in Config.in.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>